change redux composition
