+++ src/services/redux/action-types/currentIngredientTypes.js
@@ -1,0 +1,3 @@
+export const SET_CURRENT_INGREDIENT = 'SET_CURRENT_INGREDIENT',
+  DELETE_CURRENT_INGREDIENT = 'DELETE_CURRENT_INGREDIENT'
+
