+++ src/services/redux/action-types/currentIngredientsTypes.js
@@ -1,0 +1,6 @@
+export const SET_CURRENT_INGREDIENTS = 'SET_CURRENT_INGREDIENTS',
+  DELETE_CURRENT_INGREDIENTS = 'DELETE_CURRENT_INGREDIENTS',
+  CHANGE_CURRENT_ITEM_INDEX = 'CHANGE_CURRENT_ITEM_INDEX',
+  SET_ORDER_SUCCESS = 'SET_ORDER_SUCCESS',
+  SET_ORDER_REQUEST = 'SET_ORDER_REQUEST',
+  SET_ORDER_ERROR = 'SET_ORDER_ERROR'
