+++ src/services/redux/action-types/index.js
@@ -1,0 +1,5 @@
+export * from './wsActionTypes'
+export * from './userActionTypes'
+export * from './ingredientsActionTypes'
+export * from './currentIngredientsTypes'
+export * from './currentIngredientTypes'
