+++ src/services/redux/action-types/ingredientsActionTypes.js
@@ -1,0 +1,3 @@
+export const SET_INGREDIENTS_SUCCESS = 'SET_INGREDIENTS_SUCCESS',
+  SET_INGREDIENTS_ERROR = 'SET_INGREDIENTS_ERROR',
+  SET_INGREDIENTS_REQUEST = 'SET_INGREDIENTS_REQUEST'
