+++ src/services/redux/action-types/userActionTypes.js
@@ -1,0 +1,10 @@
+export const SET_USER = 'SET_USER',
+  REMOVE_USER = 'REMOVE_USER',
+  SET_USER_ERROR = 'SET_USER_ERROR',
+  SET_USER_REQUEST = 'SET_USER_REQUEST',
+  SET_SENT_EMAIL_FORGOT_PASSWORD = 'SET_SENT_EMAIL_FORGOT_PASSWORD',
+  SET_RESET_PASSWORD_STATUS = 'SET_RESET_PASSWORD_STATUS',
+  SET_INITIAL_RESET_PASSWORD_STATUS = 'SET_INITIAL_RESET_PASSWORD_STATUS',
+  CHANGE_USER_DATA = 'CHANGE_USER_DATA',
+  SET_LOAD_STATUS_TRUE = 'SET_LOAD_STATUS_TRUE',
+  SET_LOAD_STATUS_FALSE = 'SET_LOAD_STATUS_FALSE'
