+++ src/services/redux/actions/index.js
@@ -1,0 +1,5 @@
+export * from './wsActions'
+export * from './userActions'
+export * from './ingredientsAction'
+export * from './currentIngredientsActions'
+export * from './currentIngredientActions'
