+++ src/services/redux/selectors/index.js
@@ -1,0 +1,2 @@
+export * from './wsSelectors';
+export * from './getWsConnected';
